Capture 26 hours of data instead of 24
Capture the previous 26 hours of data in order to include up to 2 hours of data that is saved to the database after the start of this script (which is run daily).  We have seen cases where data was scanned 1 hour prior to the start of the script, yet it was not saved to the database until after the script started, so it was excluded from the daily extract files.  We chose 2 extra hours just to be safe.  Although this means consecutive daily extracts can have some duplicated data, that is preferable to missing data.