Change the class imported from the configparser library
As currently written the following warning is emitted when using the
scripts:
DeprecationWarning: The SafeConfigParser class has been renamed to ConfigParser in Python 3.2. This alias will be removed in future versions. Use ConfigParser directly instead.

Since this project will only support Python 3.6+ going forward it makes
sense to update to remove this warning.